# Optimized listing deployment events