Fixes #50 

This PR fixes:
- Requested events of the deployment and its replicaSets and pods in parallel.
- Filtered events from the resources of different deployments with the same labels.