@choujimmy 2019-12-06
Can you provide some comparison data before and after optimization?