@yiwei-C 2019-12-12
A Kubernetes namespace with a single deployment with 50 replicas and 260 events was used to test the **listing deployment events** performance improvement. 

4 consecutive response time was recorded as follow: 
_Original:_
```
- 9.238s
- 8.688s
- 8.697s
- 8.710s
Average Response Time: 8.833s
```
_Optimized:_
```
- 637.186ms
- 702.141ms
- 518.230ms
- 804.969ms
Average: 665.632ms
```

@choujimmy  ^^